refactor: create InsertValues class
